requirements.txt
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
contourpy==1.3.1
cycler==0.12.1
fiona==1.10.1
fonttools==4.55.3
geopandas==0.14.4
gpxpy==1.6.2
idna==3.10
kiwisolver==1.4.7
matplotlib==3.10.0
networkx==3.4.2
numpy==2.2.0
osmnx==1.2.2
packaging==24.2
pandas==2.2.3
pillow==11.0.0
prettymaps==1.0.0
pyparsing==3.2.0
pyproj==3.7.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0
requests==2.32.3
Rtree==1.3.0
Shapely==1.8.5.post1
six==1.17.0
tzdata==2024.2
urllib3==2.2.3
